HDDS-9203. Adjust secret REST endpoint to be more in line with REST paradigm
