+++ hadoop-ozone/s3gateway/src/main/java/org/apache/hadoop/ozone/s3secret/S3SecretManagementEndpoint.java
@@ -19,27 +19,36 @@
 package org.apache.hadoop.ozone.s3secret;
 
 import org.apache.hadoop.ozone.audit.S3GAction;
+import org.apache.hadoop.ozone.om.exceptions.OMException;
 import org.apache.hadoop.ozone.om.helpers.S3SecretValue;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nullable;
-import javax.ws.rs.POST;
+import javax.ws.rs.DELETE;
+import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.core.Response;
 import java.io.IOException;
 
+import static javax.ws.rs.core.Response.Status.NOT_FOUND;
+
 /**
  * Endpoint to generate and return S3 secret.
  */
-@Path("/secret/generate")
+@Path("/secret")
 @S3SecretEnabled
-public class S3SecretGenerateEndpoint extends S3SecretEndpointBase {
-  @POST
+public class S3SecretManagementEndpoint extends S3SecretEndpointBase {
+  private static final Logger LOG =
+      LoggerFactory.getLogger(S3SecretManagementEndpoint.class);
+
+  @PUT
   public Response generate() throws IOException {
     return generateInternal(null);
   }
 
-  @POST
+  @PUT
   @Path("/{username}")
   public Response generate(@PathParam("username") String username)
       throws IOException {
@@ -61,5 +70,43 @@ public class S3SecretGenerateEndpoint extends S3SecretEndpointBase {
       throws IOException {
     String actualUsername = username == null ? userNameFromRequest() : username;
     return getClient().getObjectStore().getS3Secret(actualUsername);
+  }
+
+  @DELETE
+  public Response revoke() throws IOException {
+    return revokeInternal(null);
+  }
+
+  @DELETE
+  @Path("/{username}")
+  public Response revoke(@PathParam("username") String username)
+      throws IOException {
+    return revokeInternal(username);
+  }
+
+  private Response revokeInternal(@Nullable String username)
+      throws IOException {
+    try {
+      revokeSecret(username);
+      AUDIT.logWriteSuccess(buildAuditMessageForSuccess(
+          S3GAction.REVOKE_SECRET, getAuditParameters()));
+      return Response.ok().build();
+    } catch (OMException e) {
+      AUDIT.logWriteFailure(buildAuditMessageForFailure(
+          S3GAction.REVOKE_SECRET, getAuditParameters(), e));
+      if (e.getResult() == OMException.ResultCodes.S3_SECRET_NOT_FOUND) {
+        return Response.status(NOT_FOUND.getStatusCode(),
+            OMException.ResultCodes.S3_SECRET_NOT_FOUND.toString())
+            .build();
+      } else {
+        LOG.error("Can't execute revoke secret request: ", e);
+        return Response.serverError().build();
+      }
+    }
+  }
+
+  private void revokeSecret(@Nullable String username) throws IOException {
+    String actualUsername = username == null ? userNameFromRequest() : username;
+    getClient().getObjectStore().revokeS3Secret(actualUsername);
   }
 }
